# src/main/resources/openapi/api-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Health Records API",
    "description": "API for managing patient health records",
    "contact": {
      "name": "Health Records Team",
      "url": "https://healthrecords.com",
      "email": "support@healthrecords.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0"
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Default Server URL"
    }
  ],
...
  "paths": {
...
    "/api/patients/{id}": {
      "get": {
        "tags": [
          "Patient API"
        ],
        "summary": "Get patient by ID",
        "description": "Retrieves a patient by their ID",
        "operationId": "getPatientById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the patient to retrieve",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved patient",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "404": {
            "description": "Patient not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Patient API"
        ],
        "summary": "Update patient",
        "description": "Updates an existing patient record",
        "operationId": "updatePatient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the patient to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Patient"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Patient successfully updated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "404": {
            "description": "Patient not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Patient"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Patient API"
        ],
        "summary": "Delete patient",
        "description": "Deletes a patient record by ID",
        "operationId": "deletePatient",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the patient to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Patient successfully deleted"
          },
          "404": {
            "description": "Patient not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/doctors/{id}": {
      "get": {
        "tags": [
          "doctor-controller"
        ],
        "operationId": "getDoctorById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Doctor"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "doctor-controller"
        ],
        "operationId": "updateDoctor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Doctor"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Doctor"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "doctor-controller"
        ],
        "operationId": "deleteDoctor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/diagnoses/{id}": {
      "get": {
        "tags": [
          "diagnosis-controller"
        ],
        "operationId": "getDiagnosisById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Diagnosis"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "diagnosis-controller"
        ],
        "operationId": "updateDiagnosis",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Diagnosis"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Diagnosis"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "diagnosis-controller"
        ],
        "operationId": "deleteDiagnosis",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/cohorts/{id}": {
      "get": {
        "tags": [
          "patient-cohort-selection-set-controller"
        ],
        "operationId": "getCohortById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PatientCohortSelectionSet"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "patient-cohort-selection-set-controller"
        ],
        "operationId": "updateCohort",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientCohortSelectionSet"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PatientCohortSelectionSet"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "patient-cohort-selection-set-controller"
        ],
        "operationId": "deleteCohort",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/doctors": {
      "get": {
        "tags": [
          "doctor-controller"
        ],
        "operationId": "getAllDoctors",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Doctor"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "doctor-controller"
        ],
        "operationId": "createDoctor",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Doctor"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Doctor"
                }
              }
            }
          }
        }
      }
    },
    "/api/diagnoses": {
      "get": {
        "tags": [
          "diagnosis-controller"
        ],
        "operationId": "getAllDiagnoses",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Diagnosis"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "diagnosis-controller"
        ],
        "operationId": "createDiagnosis",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Diagnosis"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Diagnosis"
                }
              }
            }
          }
        }
      }
    },
    "/api/cohorts": {
      "get": {
        "tags": [
          "patient-cohort-selection-set-controller"
        ],
        "operationId": "getAllCohorts",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientCohortSelectionSet"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "patient-cohort-selection-set-controller"
        ],
        "operationId": "createCohort",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientCohortSelectionSet"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PatientCohortSelectionSet"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/doctors/search": {
      "get": {
        "tags": [
          "doctor-controller"
        ],
        "operationId": "findDoctorsByName",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Doctor"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/cohorts/{id}/patients": {
      "get": {
        "tags": [
          "patient-cohort-selection-set-controller"
        ],
        "operationId": "getPatientsForCohort",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientDashboard"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Doctor": {
        "required": [
          "name",
          "specialty"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "maxLength": 50,
            "minLength": 4,
            "pattern": "^[a-zA-Z ]+$",
            "type": "string"
          },
          "specialty": {
            "maxLength": 50,
            "minLength": 4,
            "pattern": "^[a-zA-Z ]+$",
            "type": "string"
          }
        }
      },
      "Patient": {
        "required": [
          "dob",
          "gender",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "maxLength": 50,
            "minLength": 4,
            "pattern": "^[a-zA-Z ]+$",
            "type": "string"
          },
          "dob": {
            "type": "string",
            "format": "date"
          },
          "gender": {
            "type": "string",
            "enum": [
              "Male",
              "Female",
              "Other"
            ]
          }
        },
        "description": "Patient object to be created"
      },
      "Diagnosis": {
        "required": [
          "code",
          "description",
          "visitId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "visitId": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "PatientCohortSelectionSet": {
        "required": [
          "description",
          "name",
          "queryCriteria"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "queryCriteria": {
            "type": "string"
          }
        }
      },
...
      "PatientDashboard": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "Male",
              "Female",
              "Other"
            ]
          },
          "dateOfBirth": {
            "type": "string"
          },
          "totalVisits": {
            "type": "integer",
            "format": "int32"
          },
          "visitsLastYear": {
            "type": "integer",
            "format": "int32"
          },
          "visits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VisitSummary"
            }
          }
        }
      },
      "VisitSummary": {
        "type": "object",
        "properties": {
          "visitId": {
            "type": "integer",
            "format": "int64"
          },
          "date": {
            "type": "string"
          },
          "doctorName": {
            "type": "string"
          },
          "diagnoses": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}